renamed service.ts to RegisteredEmailAddressService.ts
