+++ src/services/api/private-get-rea/RegisteredEmailAddressService.ts
